requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-graphql-jwt==0.4.0
filelock==3.14.0
graphene==3.3
graphene-django==3.2.1
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
pillow==10.3.0
platformdirs==4.2.2
promise==2.3
psycopg2==2.9.9
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
virtualenv==20.26.2